pyproject.toml
[tool.poetry]
name = "twitter-liked-to-discordbot"
version = "0.1.0"
description = "twitter liked to discord. use twitter developer portal."
authors = ["tetsuya-k <64536338+tetsuya-ki@users.noreply.github.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.4"
"discord.py" = "^2.2.3"
python-dotenv = "^1.0.0"
pycryptodome = "^3.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
